requirements.txt
djangorestframework
django-cors-headers
markdown
django-filter
hoerapi
django-uuidfield
